+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 3.4.0 / 2014-01-25
+
+* Added dotted notation access for Hashes. See: `Lazier::Hash#enable_dotted_access`.
+* Added `Lazier::Hash#compact` and `Lazier::Hash#compact!`.
+* Added `Lazier::Object#safe_send`.
+
 ### 3.3.10 / 2013-12-02
 
 * Fixed boolean conversion.